Clean up - delete helpers as they are all moved to the store
